+++ aicontent-20240611/alibabacloud_aicontent20240611/models/__init__.py
@@ -8,6 +8,9 @@ from ._api_key_dto import ApiKeyDTO
 from ._api_key_decrypted_dto import ApiKeyDecryptedDTO
 from ._api_key_list_qry import ApiKeyListQry
 from ._api_key_update_cmd import ApiKeyUpdateCmd
+from ._billing_bill_summary_point_dto import BillingBillSummaryPointDTO
+from ._billing_bill_summary_resp_dto import BillingBillSummaryRespDTO
+from ._billing_bill_tier_dto import BillingBillTierDTO
 from ._billing_cost_breakdown_resp_dto import BillingCostBreakdownRespDTO
 from ._billing_cost_breakdown_row_dto import BillingCostBreakdownRowDTO
 from ._client_balance_dto import ClientBalanceDTO
@@ -467,6 +470,9 @@ __all__ = [
     ApiKeyDecryptedDTO,
     ApiKeyListQry,
     ApiKeyUpdateCmd,
+    BillingBillSummaryPointDTO,
+    BillingBillSummaryRespDTO,
+    BillingBillTierDTO,
     BillingCostBreakdownRespDTO,
     BillingCostBreakdownRowDTO,
     ClientBalanceDTO,