+++ aicontent-20240611/alibabacloud_aicontent20240611/models/_model_router_query_billing_cost_breakdown_request.py
@@ -7,19 +7,25 @@ from darabonba.model import DaraModel
 class ModelRouterQueryBillingCostBreakdownRequest(DaraModel):
     def __init__(
         self,
+        client_id: int = None,
         end_time: int = None,
         granularity: str = None,
         max_results: int = None,
+        model_id: int = None,
+        model_types: str = None,
         next_token: str = None,
         page: int = None,
         page_size: int = None,
         start_time: int = None,
     ):
+        self.client_id = client_id
         # This parameter is required.
         self.end_time = end_time
         # This parameter is required.
         self.granularity = granularity
         self.max_results = max_results
+        self.model_id = model_id
+        self.model_types = model_types
         self.next_token = next_token
         self.page = page
         self.page_size = page_size
@@ -34,6 +40,9 @@ class ModelRouterQueryBillingCostBreakdownRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.client_id is not None:
+            result['clientId'] = self.client_id
+
         if self.end_time is not None:
             result['endTime'] = self.end_time
 
@@ -42,6 +51,12 @@ class ModelRouterQueryBillingCostBreakdownRequest(DaraModel):
 
         if self.max_results is not None:
             result['maxResults'] = self.max_results
+
+        if self.model_id is not None:
+            result['modelId'] = self.model_id
+
+        if self.model_types is not None:
+            result['modelTypes'] = self.model_types
 
         if self.next_token is not None:
             result['nextToken'] = self.next_token
@@ -59,6 +74,9 @@ class ModelRouterQueryBillingCostBreakdownRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('clientId') is not None:
+            self.client_id = m.get('clientId')
+
         if m.get('endTime') is not None:
             self.end_time = m.get('endTime')
 
@@ -67,6 +85,12 @@ class ModelRouterQueryBillingCostBreakdownRequest(DaraModel):
 
         if m.get('maxResults') is not None:
             self.max_results = m.get('maxResults')
+
+        if m.get('modelId') is not None:
+            self.model_id = m.get('modelId')
+
+        if m.get('modelTypes') is not None:
+            self.model_types = m.get('modelTypes')
 
         if m.get('nextToken') is not None:
             self.next_token = m.get('nextToken')